requirement.txt
selenium
pyyaml
pytest-rerunfailures
pytest-html